Refactor: separate display from calculation
